// app/api/users/[id]/route.ts
import User from "@/database/user.model";
import handleError from "@/lib/handler/error";
import { NotFoundError, ValidationError } from "@/lib/http-errors";
import dbConnect from "@/lib/mongoose";
import { UserSchema } from "@/lib/validations";
import { APIErrorResponse } from "@/types/global";
import { NextResponse } from "next/server";

export async function GET(
  request: Request,
  { params }: { params: Promise<{ id: string }> },
){
    try {
        const {id} = await params;
        if(!id) throw new NotFoundError('User');
        
        await dbConnect();
        const user = await User.findById(id);

        if(!user) throw new NotFoundError('User');

        return NextResponse.json({ success: true, data: user }, { status: 200 });
        
    } catch (error) {
        return handleError(error, 'api') as unknown as APIErrorResponse;
    }
}

export async function DELETE(request:Request,{params}:{params:Promise<{id:string}>}){
    const {id} = await params;
    try {
        if(!id) throw new NotFoundError('User');

        await dbConnect();
        const deletedUser = await User.findByIdAndDelete(id);

        if(!deletedUser) throw new NotFoundError('User');

        return NextResponse.json({success:true,data:deletedUser},{status:200})
    } catch (error) {
        return handleError(error, 'api') as unknown as APIErrorResponse;
    }
}

export async function PUT(request:Request,{params}:{params:Promise<{id:string}>}){
    const {id} = await params;
    try {
        if(!id) throw new NotFoundError('User');

        await dbConnect();
        const body = await request.json();

        const validatedData = UserSchema.partial().safeParse(body);

        if(!validatedData.success){
            throw new ValidationError(validatedData.error.flatten().fieldErrors)
        }

        const updatedUser = await User.findByIdAndUpdate(id,validatedData,{new:true});

        if(!updatedUser) throw new NotFoundError('User');

        return NextResponse.json({success:true,data:updatedUser},{status:200})
    } catch (error) {
        return handleError(error, 'api') as unknown as APIErrorResponse;
    }
}